requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
openai
streamlit
requests
python-dotenv
